requirements.txt
django
python-dotenv
psycopg2
psycopg2-binary
djangorestframework
drf-yasg
djangorestframework_simplejwt